Rename module filename.
