+++ crates/typst-library/src/layout/mod.rs
@@ -4,7 +4,6 @@ mod abs;
 mod align;
 mod angle;
 mod axes;
-mod cellgrid;
 mod columns;
 mod container;
 mod corners;
@@ -24,6 +23,7 @@ mod pad;
 mod page;
 mod place;
 mod point;
+mod raster;
 mod ratio;
 mod regions;
 mod rel;
@@ -38,7 +38,6 @@ pub use self::abs::*;
 pub use self::align::*;
 pub use self::angle::*;
 pub use self::axes::*;
-pub use self::cellgrid::*;
 pub use self::columns::*;
 pub use self::container::*;
 pub use self::corners::*;
@@ -56,6 +55,7 @@ pub use self::pad::*;
 pub use self::page::*;
 pub use self::place::*;
 pub use self::point::*;
+pub use self::raster::*;
 pub use self::ratio::*;
 pub use self::regions::*;
 pub use self::rel::*;
